apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for AI Codebase Onboarding System (scaffold)"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic-settings>=2.4",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "pgvector>=0.3.6",
  "psycopg[binary]>=3.2",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "pytest>=8.0",
  "httpx>=0.27",
]

[dependency-groups]
dev = [
  "ruff>=0.6",
  "pytest>=8.0",
  "httpx>=0.27",
]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["alembic/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

